Trim DB Storage
Trim the Lat/Lon fields to the bare minimum needed. Latitude %2.4f and
Longitude %3.4f.

+++ db/geodata.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS `locations` (
   `locid` mediumint(8) unsigned NOT NULL AUTO_INCREMENT,
-  `latitude` float(8,4) NOT NULL,
-  `longitude` float(8,4) NOT NULL,
+  `latitude` float(6,4) NOT NULL,
+  `longitude` float(7,4) NOT NULL,
   `ipcount` int(10) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`locid`),
   KEY `latlon` (`latitude`,`longitude`,`ipcount`)
